Add badge support
This returns list of user badges from the UserDao class.
It adds the BadgeDao class and the badge-list template. BadgeDao
is used to access the badges DB table. The badge-list template
is used for displaying all badges in the system and how to
attain them.
It updates the routing in index.php to handle the badge-list page
It updates the user-profile template to link to the badge list.
closes #11