Improve spacing between listed tasks in a list
To achieve this, I followed the principle of not relying on CSS
classes to apply a styling.

A list of tasks is now a set of heading and paragraph tags within
a div with ID of "tasks".

Previously, each listed task was in its own div, with a class of
.task.

It's better, it seems to either user IDs, base HTML5 elements, or
use relationship-based CSS3 selectors.

The implementation doesn't work in current versions of IE, from
what I understand. But I'm not too worried about that, given the
fact that it doesn't break the site if this CSS styling is not
applied.