Return user to previous page after registering
This commit returns users to their previous page after registering.
The user's session variable must be set while they are not logged
in to track what page they were on. When they register, this value
is read and, if they register and log in without error, they are
returned to their previous page. closes #34. It also fixes an issue
on the home screen where the user profile button got displayed
even when not logged in.