Add document preview support
Adds document preview support before the user downloads a file.
Uses google docs to display a preview of the file. Note that this
commit introduces a new htaccess file in the uploads folder which
gives permissions to anyone trying to access the file. This must
be refined to the specific ver-0 folders that are being accessed.
Starts work on issue #45