Several updates to the tag subscription mechanism.
Changes the position of the subscribe/unsubscribe button on the
tag page. Fix problem related to popular tags list being pushed into
the middle of the screen (fixes #47). Adds an x to the tags in the
subscribed list which allows the user to unsubscribe at any time
(closes #48). Repositions subscribed tags in list to be more readable.
Updates CSS related to tags, changes color of unsubscribe x when
hovered over. Removes old code in index.php that used to handle
tag subscriptions. Creates a new generalised function for handling
tag subscription that redirects the user to their previous page when
finished. Makes the subscribe button appear on the tag page even when
no tags were in the subscription list (closes #51).