Add a python script for calculating the scores for user-task pairs
This is the first iteration of a python script that calculates
user-task scores. This score is used to organise the stream on the
home page for users who are logged in. NOTE: There is a big error
in the MySQL connect, namely that I couldn't connect to MySQL
correctly on my testing PC and had to hard code the unix_socket
location. There must be some external python file where this can be
set. So far it just gives a higher score to tasks with a lot of
tags that the user has subscribed to. See issue #16. Must also set
it to run automatically