Update the SQL functions in calculate_scores.py
This commit updates the python script to use the DB config from
conf.ini. It also fixes an issue where the MySQL socket file
could not be found. It seems that, for some systems, the python
module was not reading the my.cnf file and was assuming the
socket was in its default location. This update should make the
script run accross all servers. This closes #54