Replace language selector with drop down list
Adds a file filled with languages and an entry in conf.ini which
references it. Adds a function to the Languages class for getting
a list of all the languages in the system (Consider adding langs
from the DB as well). Update task describe page so that it uses
drop down lists to select the language. #103