Add middleware and update old middleware
Some of the old middleware prevented Org Members from altering task
details for tasks that belong to the Org but were not claimed by
that org member. All these have been replaced by new middleware
for authenticating users are members of the Organisation that
uploaded the file they are altering. Closes #220