Allow users to subscribe for notifications from tasks
Updates PDOWrapper's call function to take an array or a string
to define the args. Adds new functions to UserDao for email
notification subscription and management. Add stored procedures
for handling the new functions. Adds flash messaging support to
the client dashboard. Adds a new column to the client dashboard
for task notification subscription. The new column consists of
buttons that allow users to either follow a specific task or ignore
it. Adds flash messaging support to the task view page. Adds button
to task view page to allow users to subscribe to task
notifications. Alters wording in middleware. Adds post handling
for the task view page. Adds POST handling for the client
dashboard. Closes #148