Allow logged in users to create an organisation.
Add a button to the client dashboard which allows users to create
an organisation. Organisations are identified by their names and
so the system will not allow Organisations with duplicate names.
Adds route to the Org Route Handler, adds template file for org
creation page, alters sql schema to allow orgs to be created when
only a name is supplied, adds to the OrgDao find function, Adds
button to the client dashboard that links to the new page.
Closes #147