Fix UserTaskScore Requests to use protobufs
To comply with changes in the back end brought in by the
PluginScheduler this commit alters the UserTaskScore request
to use the same proto as the back end. An empty proto calculates
the score for all tasks in the system. Alternatively, set the
task ID and only that task will get scored.