Add check if review has been submitted on review page
The page now checks to see if the user has already submitted a
review for that file. If he/she has then a message is displayed
and informing them and the fields are auto populated with their
review data. Adds a number of functions to the API and UI for
reviews.