Update claimed tasks page and add user task reviews
Update claimed tasks page so that it runs on dart. It now only
loads the required tasks for the page being viewed and changes
pages with requiring a refresh of the url. It also adds a link to
task reviews on the claimed page. This link takes the user to a page
where they can see all the reviews for that task. Part of #432
Closes #497