Add support for task reviews to orgs
Refactor the task complete page. Now it has one button to download
the output file, one buttong to go to a new page where a review
can be subbmitted and one button to got to a new page which displays
all reviews for that task (including input and output file reviews).
Refactors the task review template so it can be used on numerous
pages. Closes #498