Added tool for merging site translations
Adds a tool that takes 3 files, the original english strings file,
the current version of a specific language strings file and a
file containing all the new strings for that language. It then
parses all the files and merges the third file with the second.
Needs an update to merge new strings. Part of #740