Fix error in archive task.
There were 2 problems here. First, a task must have a file
associated with it. If you try to delete a task with no file
info it will fail. Secondly, the model factory was broken for
archived task generation. It was using incorrect field names which
caused the creation of partial objects. Closes #751