Altered the DataAccessObjects in the api so that get functions for each of our models can now return a single object as well as an array of objects.  For example, if searching by a unique ID only a single object should be returned and not an array containing one object.
The previous get functions were replaced with two functions.  One that returns multiple objects and one that returns single objects (eg. getBadges and getBadge).

Some functions required one or both of two arguments, such a the getUser function that can accept an id and/or email.  This was used in the newly created getUser method.

All previous methods, that return arrays had to be renamed to reflect that they are now plurals (eg.  getUsers was previously getUser).  This necesitated a change accross multiple files that previously called the singular named function but must now call the plural version.

Some of the api > lib files were calling the singular get when they should now be calling the plural version.
The api > lib files that should be calling a singular object were also assigning their value as the first index in an array of returned objects, which is no longer the case as they now call the singular methods.  Removed a line of code from each of these as it was now redundant.

Fixed a naming issue in one method in the ui Middleware.class.php that was causing an error.

Removed a call to getTaskStatus in the UserDao.class.php file (api) as it is no longer used.  The status of a task is currently being returned by the stored procedure.