Fixes #1152
API: When try to create an org with an existing name, the following code always returns the first org (alphabetically sorted)
$org = DAO\OrganisationDao::insertAndUpdate($data);
Prevented this error by encapsulating the org creation part under ELSE statment of the previous IF statement. Also removed redundant call for acceptMemRequest as it is being done as a part of OrgAdmin Procedure. Also commented the part where orgs can get accidently deleted by any unforseen bugs in the code .

UI/DAO Change: remove redundant code. OrgAdmin is created as a part of the API call.

Schema change: Ensure that the newly created organisation is always returned. This is possible because the organisation names have to be unique by the design.