Only send notification once (within a period of 24 hours) for deadline passed
This addresses https://github.com/TheRosettaFoundation/SOLAS-Match-Backend/issues/45
Notes...

The actual period is 24.5 hours to ensure none are missed.
This might cause some to be duplicated but this is not really an issue
(also almost no deadlines with be for around the time of 7 am so there should
be almost no duplicates).

Notifications might be missed if the background is down.

The above change relies on the deadline checker being run once a day:-
/repo/SOLAS-Match-Backend/schedule.xml
    <task startTime="07:05">
        <interval days="1" hours="0" mins="0" secs="0"></interval>
        <message exchange="SOLAS_MATCH" topic="tasks.deadline.check">
            DeadlineCheckRequest
        </message>
    </task>